pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Shibyan"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13" 
dependencies = [
    "nemo-toolkit[all] (>=2.3.0,<3.0.0)",
    "kokoro (>=0.9.4,<0.10.0)",
    "sounddevice (>=0.5.2,<0.6.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "langchain-core (>=0.3.60,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "websockets (>=15.0.1,<16.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
